Simplify environment and XDG path helpers
