+++ docking/applets/camshield/helper.py
@@ -31,6 +31,8 @@ from contextlib import suppress
 from dataclasses import dataclass
 from pathlib import Path
 
+from docking.core.paths import ensure_dir
+
 STATE_DIR = Path("/var/lib/docking/camshield")
 STATE_FILE = STATE_DIR / "devices.json"
 DEV_ROOT = Path("/dev")
@@ -73,7 +75,7 @@ def lock_devices(
         print("No camera devices found.", file=sys.stderr)
         return 1
 
-    state_dir.mkdir(parents=True, exist_ok=True)
+    ensure_dir(state_dir)
     state = _read_state(state_dir=state_dir)
     changed = 0
     for device in devices:
@@ -157,7 +159,7 @@ def _save_acl_snapshot(*, device: Path, state_dir: Path) -> Path | None:
         return None
 
     acl_dir = state_dir / "acl"
-    acl_dir.mkdir(parents=True, exist_ok=True)
+    ensure_dir(acl_dir)
     acl_file = acl_dir / f"{device.name}.acl"
     try:
         with acl_file.open("w", encoding="utf-8") as handle:
